+++ src/palettes.cpp
@@ -1,0 +1,21 @@
+#include <palettes.h>
+
+extern const TProgmemPalette16 purpleAndBlackPalette PROGMEM =
+{
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Purple,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Purple,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Purple,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Black,
+  CRGB::Purple
+};
