+++ src/palettes.h
@@ -1,0 +1,8 @@
+#ifndef FLOWER_PALETTES_H
+#define FLOWER_PALETTES_H
+
+#include <FastLED.h>
+
+extern const TProgmemPalette16 purpleAndBlackPalette PROGMEM;
+
+#endif  // FLOWER_PALETTES_H
