+++ src/patterns.h
@@ -1,0 +1,12 @@
+#ifndef FLOWER_PATTERNS_H
+#define FLOWER_PATTERNS_H
+
+#include <globals.h>
+
+void rainbowPetals();
+void rainbowFlower();
+void insideOut();
+void sineMove();
+void plasma();
+
+#endif  // FLOWER_PATTERNS_H
